+++ spec/presenters/budget_presenter_spec.rb
@@ -3,6 +3,12 @@ require "rails_helper"
 RSpec.describe BudgetPresenter do
   let(:budget) { build_stubbed(:budget, financial_year: 2020, value: "20") }
 
+  describe "#budget_type" do
+    it "returns the name of the budget type" do
+      expect(described_class.new(budget).budget_type).to eq("Direct (Newton fund)")
+    end
+  end
+
   describe "#iati_type" do
     it "returns the name of the IATI budget type" do
       expect(described_class.new(budget).iati_type).to eq("Original")