+++ CHANGELOG.md
@@ -579,6 +579,7 @@
 - Add supporting hint text about the soft limits to title and description
 - Change channel of delivery code to radio buttons
 - Collect transferred and external budget types
+- Show the type of budget in the application
 
 [unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-42...HEAD
 [release-42]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-41...release-42